refactor: separate commonly used objects to a utility package
