+++ producer/config.py
@@ -1,27 +1,7 @@
-from dataclasses import dataclass
 import argparse
 
 from confluent_kafka import Producer
 from confluent_kafka.schema_registry import SchemaRegistryClient
-
-
-@dataclass
-class Topics:
-    TweetsTopic: str = "Model.Tweets.1"
-    UsersTopic: str = "Model.Users.1"
-    CommentsTopic: str = "Model.Comments.1"
-    TweetLikesTopic: str = "Model.TweetLikes.1"
-    UserFollowsTopic: str = "Model.UserFollows.1"
-
-
-# Dictionary of topics and their producing probability.
-TOPICS_TO_PRODUCING_PROBABILITY = {
-    Topics.TweetsTopic: 0.3,
-    Topics.UsersTopic: 0.2,
-    Topics.CommentsTopic: 0.2,
-    Topics.TweetLikesTopic: 0.1,
-    Topics.UserFollowsTopic: 0.1,
-}
 
 
 class ClientGenerator:
